refactor: useImageUpload 위치 변경
- useImageUpload 커스텀 훅을 common 폴더로 이동
